requirements.txt
altgraph==0.17.2
ansible==4.3.0
ansible-core==2.11.3
black==21.9b0
bluepy==1.3.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.3
cryptography==3.4.7
idna==3.2
Jinja2==3.0.1
JPype1==1.3.0
kafka-python==2.0.2
kafka-tools==0.1.9
kazoo==2.8.0
libxml2-python==2.9.10
lxml==4.6.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
pex==2.1.52
platformdirs==2.4.0
pycparser==2.20
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pyparsing==2.4.7
PyYAML==5.4.1
regex==2021.10.23
requests==2.26.0
resolvelib==0.5.4
six==1.16.0
tabulate==0.8.9
tomli==1.2.2
typing-extensions==3.10.0.2
urllib3==1.26.6
